requirements.txt
moviepy
pillow
numpy
tqdm
pytest
librosa
